The color "grey" should be spelled "gray"
Is "gray" the version more commonly used in American English, and "grey" the version more commonly used in British English? A-gray for America and E-grey for England I guess. Is it weird that I as an American used "grey" more often? But the Racket-Color-Language uses "gray", so I'm not referring to the English word, but the Racket-Color-Language word.